Update admin applications url
