+++ pages/admin/applications/finalised-for-review.tsx
@@ -1,12 +1,12 @@
 import { useEffect, useState } from 'react'
-import { AdminPortalData } from '../../classes/admin_portal_data'
-import { ApplicationData } from '../../classes/application_data'
-import ApplicationsTable from '../../components/Admin/ApplicationsTable'
-import Loading from '../../components/Loading'
-import requireAuth from '../../components/requireAuth'
-import Roles from '../../constants/roles'
-import AdminLayout from '../../layouts/admin/admin-layout'
-import { getApplicationsWithGivenStatus } from '../api/getApplicationsResponse'
+import { AdminPortalData } from '../../../classes/admin_portal_data'
+import { ApplicationData } from '../../../classes/application_data'
+import ApplicationsTable from '../../../components/Admin/ApplicationsTable'
+import Loading from '../../../components/Loading'
+import requireAuth from '../../../components/requireAuth'
+import Roles from '../../../constants/roles'
+import AdminLayout from '../../../layouts/admin/admin-layout'
+import { getApplicationsWithGivenStatus } from '../../api/getApplicationsResponse'
 
 type Applications = {
   [key: string]: {
@@ -15,13 +15,13 @@ type Applications = {
   }
 }
 
-function FinalisedApplicationsForInterview() {
+function FinalisedApplicationsForReview() {
   const [applications, setApplications] = useState<Applications>()
   const [changeOccured, setChangeOccured] = useState<boolean>(false)
   const [pageReady, setPageReady] = useState<boolean>(false)
 
   useEffect(() => {
-    getApplicationsWithGivenStatus(4)
+    getApplicationsWithGivenStatus(2)
       .then((data) => {
         setApplications(data)
       })
@@ -38,10 +38,10 @@ function FinalisedApplicationsForInterview() {
           setChangeOccured={setChangeOccured}
         />
       ) : (
-        <Loading message="Loading your applications!" />
+        <Loading message="Loading finalised applications for review!" />
       )}
     </AdminLayout>
   )
 }
 
-export default requireAuth(FinalisedApplicationsForInterview, Roles.ADMIN)
+export default requireAuth(FinalisedApplicationsForReview, Roles.ADMIN)
